# internal/adapter/driver/api/dto/create_order_dto.go
package dto

type CreateOrderDTO struct {
	Items       []*CreateOrderItemDTO `json:"items"`
	CustomerCPF string                `json:"cpf"`
}

type CreateOrderItemDTO struct {
	ProductID string `json:"productid"`
	Quantity  int    `json:"quantity"`
}